refactor: Reconfigure JSON translation functionalities
The translation functionalities have been reconfigured to support blocks of data and exclusion parameters. The restructuring included creating a new "JsonTranslator" class in the "json_translator.py" file and updating the "GoogleTranslate" and "MicrosoftTranslator" classes. Furthermore, some repetitive code was removed which accelerated the translation process.
