+++ src/_main_/utils/translation/translator/providers/microsoft_translator.py
@@ -1,7 +1,7 @@
 import os, requests, uuid, json
 
 class MicrosoftTranslator:
-    MAX_CHAR_LIMIT = 10000
+    MAX_TEXT_SIZE = 10000
 
     def __init__(self):
         self.__name__ = "MicrosoftTranslator"
